# executor: create PipelinedWindowExec