--- a/cmd/explaintest/r/select.result
+++ b/cmd/explaintest/r/select.result
@@ -468,7 +468,7 @@ PRIMARY KEY (`id`)
 explain format = 'brief' select row_number() over( partition by i ) - x as rnk from t;
 id	estRows	task	access object	operator info
 Projection	10000.00	root		minus(Column#5, test.t.x)->Column#7
-└─Window	10000.00	root		row_number()->Column#5 over(partition by test.t.i)
+└─Window	10000.00	root		row_number()->Column#5 over(partition by test.t.i rows between current row and current row)
   └─Sort	10000.00	root		test.t.i
     └─TableReader	10000.00	root		data:TableRangeScan
       └─TableRangeScan	10000.00	cop[tikv]	table:t	range:[0,+inf], keep order:false, stats:pseudo