--- a/executor/aggfuncs/func_bitfuncs.go
+++ b/executor/aggfuncs/func_bitfuncs.go
@@ -91,10 +91,12 @@ func (e *bitXorUint64) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup 
 	return memDelta, nil
 }
 
-func (e *bitXorUint64) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &bitXorUint64{}
+
+func (e *bitXorUint64) Slide(sctx sessionctx.Context, getRow func(uint64) chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	p := (*partialResult4BitFunc)(pr)
 	for i := uint64(0); i < shiftStart; i++ {
-		inputValue, isNull, err := e.args[0].EvalInt(sctx, rows[lastStart+i])
+		inputValue, isNull, err := e.args[0].EvalInt(sctx, getRow(lastStart+i))
 		if err != nil {
 			return err
 		}
@@ -104,7 +106,7 @@ func (e *bitXorUint64) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStar
 		*p ^= uint64(inputValue)
 	}
 	for i := uint64(0); i < shiftEnd; i++ {
-		inputValue, isNull, err := e.args[0].EvalInt(sctx, rows[lastEnd+i])
+		inputValue, isNull, err := e.args[0].EvalInt(sctx, getRow(lastEnd+i))
 		if err != nil {
 			return err
 		}