--- a/executor/aggfuncs/func_count.go
+++ b/executor/aggfuncs/func_count.go
@@ -68,10 +68,12 @@ func (e *countOriginal4Int) UpdatePartialResult(sctx sessionctx.Context, rowsInG
 	return 0, nil
 }
 
-func (e *countOriginal4Int) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &countOriginal4Int{}
+
+func (e *countOriginal4Int) Slide(sctx sessionctx.Context, getRow func(uint64)chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	p := (*partialResult4Count)(pr)
 	for i := uint64(0); i < shiftStart; i++ {
-		_, isNull, err := e.args[0].EvalInt(sctx, rows[lastStart+i])
+		_, isNull, err := e.args[0].EvalInt(sctx, getRow(lastStart+i))
 		if err != nil {
 			return err
 		}
@@ -81,7 +83,7 @@ func (e *countOriginal4Int) Slide(sctx sessionctx.Context, rows []chunk.Row, las
 		*p--
 	}
 	for i := uint64(0); i < shiftEnd; i++ {
-		_, isNull, err := e.args[0].EvalInt(sctx, rows[lastEnd+i])
+		_, isNull, err := e.args[0].EvalInt(sctx, getRow(lastEnd+i))
 		if err != nil {
 			return err
 		}