--- a/executor/aggfuncs/func_count.go
+++ b/executor/aggfuncs/func_count.go
@@ -117,10 +117,12 @@ func (e *countOriginal4Real) UpdatePartialResult(sctx sessionctx.Context, rowsIn
 	return 0, nil
 }
 
-func (e *countOriginal4Real) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &countOriginal4Real{}
+
+func (e *countOriginal4Real) Slide(sctx sessionctx.Context, getRow func(uint64) chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	p := (*partialResult4Count)(pr)
 	for i := uint64(0); i < shiftStart; i++ {
-		_, isNull, err := e.args[0].EvalReal(sctx, rows[lastStart+i])
+		_, isNull, err := e.args[0].EvalReal(sctx, getRow(lastStart+i))
 		if err != nil {
 			return err
 		}
@@ -130,7 +132,7 @@ func (e *countOriginal4Real) Slide(sctx sessionctx.Context, rows []chunk.Row, la
 		*p--
 	}
 	for i := uint64(0); i < shiftEnd; i++ {
-		_, isNull, err := e.args[0].EvalReal(sctx, rows[lastEnd+i])
+		_, isNull, err := e.args[0].EvalReal(sctx, getRow(lastEnd+i))
 		if err != nil {
 			return err
 		}
@@ -164,10 +166,12 @@ func (e *countOriginal4Decimal) UpdatePartialResult(sctx sessionctx.Context, row
 	return 0, nil
 }
 
-func (e *countOriginal4Decimal) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &countOriginal4Decimal{}
+
+func (e *countOriginal4Decimal) Slide(sctx sessionctx.Context, getRow func(uint64) chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	p := (*partialResult4Count)(pr)
 	for i := uint64(0); i < shiftStart; i++ {
-		_, isNull, err := e.args[0].EvalDecimal(sctx, rows[lastStart+i])
+		_, isNull, err := e.args[0].EvalDecimal(sctx, getRow(lastStart+i))
 		if err != nil {
 			return err
 		}
@@ -177,7 +181,7 @@ func (e *countOriginal4Decimal) Slide(sctx sessionctx.Context, rows []chunk.Row,
 		*p--
 	}
 	for i := uint64(0); i < shiftEnd; i++ {
-		_, isNull, err := e.args[0].EvalDecimal(sctx, rows[lastEnd+i])
+		_, isNull, err := e.args[0].EvalDecimal(sctx, getRow(lastEnd+i))
 		if err != nil {
 			return err
 		}
@@ -211,10 +215,12 @@ func (e *countOriginal4Time) UpdatePartialResult(sctx sessionctx.Context, rowsIn
 	return 0, nil
 }
 
-func (e *countOriginal4Time) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &countOriginal4Time{}
+
+func (e *countOriginal4Time) Slide(sctx sessionctx.Context, getRow func(uint64) chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	p := (*partialResult4Count)(pr)
 	for i := uint64(0); i < shiftStart; i++ {
-		_, isNull, err := e.args[0].EvalTime(sctx, rows[lastStart+i])
+		_, isNull, err := e.args[0].EvalTime(sctx, getRow(lastStart+i))
 		if err != nil {
 			return err
 		}
@@ -224,7 +230,7 @@ func (e *countOriginal4Time) Slide(sctx sessionctx.Context, rows []chunk.Row, la
 		*p--
 	}
 	for i := uint64(0); i < shiftEnd; i++ {
-		_, isNull, err := e.args[0].EvalTime(sctx, rows[lastEnd+i])
+		_, isNull, err := e.args[0].EvalTime(sctx, getRow(lastEnd+i))
 		if err != nil {
 			return err
 		}
@@ -258,10 +264,12 @@ func (e *countOriginal4Duration) UpdatePartialResult(sctx sessionctx.Context, ro
 	return 0, nil
 }
 
-func (e *countOriginal4Duration) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &countOriginal4Duration{}
+
+func (e *countOriginal4Duration) Slide(sctx sessionctx.Context, getRow func(uint64) chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	p := (*partialResult4Count)(pr)
 	for i := uint64(0); i < shiftStart; i++ {
-		_, isNull, err := e.args[0].EvalDuration(sctx, rows[lastStart+i])
+		_, isNull, err := e.args[0].EvalDuration(sctx, getRow(lastStart+i))
 		if err != nil {
 			return err
 		}
@@ -271,7 +279,7 @@ func (e *countOriginal4Duration) Slide(sctx sessionctx.Context, rows []chunk.Row
 		*p--
 	}
 	for i := uint64(0); i < shiftEnd; i++ {
-		_, isNull, err := e.args[0].EvalDuration(sctx, rows[lastEnd+i])
+		_, isNull, err := e.args[0].EvalDuration(sctx, getRow(lastEnd+i))
 		if err != nil {
 			return err
 		}
@@ -305,10 +313,12 @@ func (e *countOriginal4JSON) UpdatePartialResult(sctx sessionctx.Context, rowsIn
 	return 0, nil
 }
 
-func (e *countOriginal4JSON) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &countOriginal4JSON{}
+
+func (e *countOriginal4JSON) Slide(sctx sessionctx.Context, getRow func(uint64) chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	p := (*partialResult4Count)(pr)
 	for i := uint64(0); i < shiftStart; i++ {
-		_, isNull, err := e.args[0].EvalJSON(sctx, rows[lastStart+i])
+		_, isNull, err := e.args[0].EvalJSON(sctx, getRow(lastStart+i))
 		if err != nil {
 			return err
 		}
@@ -318,7 +328,7 @@ func (e *countOriginal4JSON) Slide(sctx sessionctx.Context, rows []chunk.Row, la
 		*p--
 	}
 	for i := uint64(0); i < shiftEnd; i++ {
-		_, isNull, err := e.args[0].EvalJSON(sctx, rows[lastEnd+i])
+		_, isNull, err := e.args[0].EvalJSON(sctx, getRow(lastEnd+i))
 		if err != nil {
 			return err
 		}
@@ -352,10 +362,12 @@ func (e *countOriginal4String) UpdatePartialResult(sctx sessionctx.Context, rows
 	return 0, nil
 }
 
-func (e *countOriginal4String) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &countOriginal4String{}
+
+func (e *countOriginal4String) Slide(sctx sessionctx.Context, getRow func(uint64) chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	p := (*partialResult4Count)(pr)
 	for i := uint64(0); i < shiftStart; i++ {
-		_, isNull, err := e.args[0].EvalString(sctx, rows[lastStart+i])
+		_, isNull, err := e.args[0].EvalString(sctx, getRow(lastStart+i))
 		if err != nil {
 			return err
 		}
@@ -365,7 +377,7 @@ func (e *countOriginal4String) Slide(sctx sessionctx.Context, rows []chunk.Row, 
 		*p--
 	}
 	for i := uint64(0); i < shiftEnd; i++ {
-		_, isNull, err := e.args[0].EvalString(sctx, rows[lastEnd+i])
+		_, isNull, err := e.args[0].EvalString(sctx, getRow(lastEnd+i))
 		if err != nil {
 			return err
 		}