--- a/executor/aggfuncs/row_number.go
+++ b/executor/aggfuncs/row_number.go
@@ -53,6 +53,8 @@ func (rn *rowNumber) AppendFinalResult2Chunk(sctx sessionctx.Context, pr Partial
 	return nil
 }
 
-func (rn *rowNumber) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+var _ SlidingWindowAggFunc = &rowNumber{}
+
+func (rn *rowNumber) Slide(sctx sessionctx.Context, getRow func(uint64)chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
 	return nil
 }