… executor/benchmark_test.go
… executor/pipelined_window.go
… executor/window.go
… executor/window_test.go
… expression/aggregation/window_func.go
… planner/core/logical_plan_builder.go
… sessionctx/variable/session.go
… sessionctx/variable/sysvar.go
… sessionctx/variable/tidb_vars.go