<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?


Problem Summary:

### What is changed and how it works?

Proposal: [see design doc](https://github.com/ichn-hu/tidb/blob/add-pw-dc/docs/design/2021-03-01-pipelined-window-functions.md) <!-- REMOVE this line if not applicable -->

What's Changed:

How it Works:

### Related changes

- PR to update `pingcap/docs`/`pingcap/docs-cn`:
- Need to cherry-pick to the release branch

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- No code

Side effects


### Release note <!-- bugfixes or new feature need a release note -->

- No release note