@ichn-hu 2021-03-01
> > @Tjianke I realized I can just turn the doc to open to web, so here you are: [doc](https://docs.google.com/document/d/e/2PACX-1vR1LMxbkvR2doGJudMl_jbFTP4xbvwQ2EczJMDu5HsVpfOZ67RJZw32P9Y0FmCUvl9-zRd_S0N96ymO/pub). The link in the PR description is also updated.
> 
> I'm afraid the link is still not generally available, prompting `corresponding rights needed`.

Sooorry, please see the design doc PR: https://github.com/ichn-hu/tidb/blob/add-pw-dc/docs/design/2021-03-01-pipelined-window-functions.md