@ichn-hu 2021-05-31
the benchstat result is

there is no significant regression on all window functions related benchmark

```
name                                                                                                                                        old time/op    new time/op    delta
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:1000,_sorted:false,_concurrency:1,_)-24                                 443µs ± 0%     474µs ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:1000,_sorted:false,_concurrency:2,_)-24                                1.39ms ± 0%    1.44ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:1000,_sorted:false,_concurrency:4,_)-24                                3.80ms ± 0%    3.58ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:1000,_sorted:false,_concurrency:1,_)-24                                597µs ± 0%     623µs ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:1000,_sorted:false,_concurrency:2,_)-24                               1.48ms ± 0%    1.50ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:1000,_sorted:false,_concurrency:4,_)-24                               4.53ms ± 0%    4.45ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:1000,_sorted:false,_concurrency:1,_)-24                              948µs ± 0%     999µs ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:1000,_sorted:false,_concurrency:2,_)-24                             1.54ms ± 0%    1.56ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:1000,_sorted:false,_concurrency:4,_)-24                             4.71ms ± 0%    4.46ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:100000,_sorted:false,_concurrency:1,_)-24                              34.8ms ± 0%    36.1ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:100000,_sorted:false,_concurrency:2,_)-24                              53.4ms ± 0%    55.8ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:100000,_sorted:false,_concurrency:4,_)-24                              51.8ms ± 0%    54.1ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:100000,_sorted:false,_concurrency:1,_)-24                             53.5ms ± 0%    57.3ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:100000,_sorted:false,_concurrency:2,_)-24                             62.4ms ± 0%    50.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:100000,_sorted:false,_concurrency:4,_)-24                             36.6ms ± 0%    37.0ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24                           93.4ms ± 0%    88.8ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:2,_)-24                           68.9ms ± 0%    70.6ms ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24                           47.9ms ± 0%    47.5ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctions/(func:lag,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24                             97.3ms ± 0%    84.3ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctions/(func:lag,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24                             48.6ms ± 0%    47.6ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24             99.0ms ± 0%    92.5ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:2,_)-24             73.2ms ± 0%    74.2ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:3,_)-24             56.9ms ± 0%    56.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24             51.3ms ± 0%    50.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:5,_)-24             46.0ms ± 0%    45.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:6,_)-24             42.3ms ± 0%    42.3ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24              141ms ± 0%     135ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:2,_)-24              100ms ± 0%      98ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:3,_)-24             76.4ms ± 0%    72.8ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24             66.9ms ± 0%    62.8ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:5,_)-24             57.3ms ± 0%    56.2ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:6,_)-24             52.6ms ± 0%    51.6ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:1,_)-24              13.9ms ± 0%    14.4ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:2,_)-24              28.0ms ± 0%    28.2ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:3,_)-24              28.3ms ± 0%    28.4ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:4,_)-24              28.2ms ± 0%    28.3ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:5,_)-24              28.3ms ± 0%    28.5ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:6,_)-24              28.4ms ± 0%    28.4ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:1,_)-24              55.1ms ± 0%    50.1ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:2,_)-24              44.0ms ± 0%    40.6ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:3,_)-24              37.3ms ± 0%    34.0ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:4,_)-24              34.0ms ± 0%    32.0ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:5,_)-24              32.7ms ± 0%    31.3ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:6,_)-24              32.8ms ± 0%    31.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24                91.0ms ± 0%    91.2ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:2,_)-24                68.7ms ± 0%    72.3ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:3,_)-24                55.2ms ± 0%    53.8ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24                47.4ms ± 0%    47.2ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:5,_)-24                42.3ms ± 0%    43.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:6,_)-24                40.5ms ± 0%    41.0ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24                 117ms ± 0%     113ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:2,_)-24                77.2ms ± 0%    80.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:3,_)-24                61.8ms ± 0%    63.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24                55.1ms ± 0%    54.4ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:5,_)-24                49.7ms ± 0%    48.4ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:6,_)-24                44.4ms ± 0%    46.7ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:1,_)-24                 9.45ms ± 0%    9.90ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:2,_)-24                 27.3ms ± 0%    27.8ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:3,_)-24                 26.9ms ± 0%    27.2ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:4,_)-24                 27.6ms ± 0%    27.3ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:5,_)-24                 27.5ms ± 0%    27.9ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:true,_concurrency:6,_)-24                 27.3ms ± 0%    27.8ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:1,_)-24                 28.4ms ± 0%    30.4ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:2,_)-24                 29.8ms ± 0%    30.9ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:3,_)-24                 29.8ms ± 0%    29.8ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:4,_)-24                 29.9ms ± 0%    29.9ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:5,_)-24                 30.0ms ± 0%    29.6ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:bit_xor,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:true,_concurrency:6,_)-24                 29.9ms ± 0%    30.0ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsAggWindowProcessorAboutFrame/(func:max,_aggColType:double,_numFunc:1,_ndv:10,_rows:10000,_sorted:false,_concurrency:1,_)-24    8.40ms ± 0%    6.46ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithSlidingWindow/(func:sum,_aggColType:float,_numFunc:1,_ndv:100,_rows:100000,_sorted:true,_concurrency:1,_)-24               18.3ms ± 0%    19.9ms ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithSlidingWindow/(func:sum,_aggColType:decimal(10,0),_numFunc:1,_ndv:100,_rows:100000,_sorted:true,_concurrency:1,_)-24       35.6ms ± 0%    38.8ms ± 0%   ~     (p=1.000 n=1+1)

name                                                                                                                                        old alloc/op   new alloc/op   delta
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:1000,_sorted:false,_concurrency:1,_)-24                                 279kB ± 0%     279kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:1000,_sorted:false,_concurrency:2,_)-24                                 593kB ± 0%     569kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:1000,_sorted:false,_concurrency:4,_)-24                                 590kB ± 0%     614kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:1000,_sorted:false,_concurrency:1,_)-24                                287kB ± 0%     279kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:1000,_sorted:false,_concurrency:2,_)-24                                560kB ± 0%     549kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:1000,_sorted:false,_concurrency:4,_)-24                                592kB ± 0%     558kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:1000,_sorted:false,_concurrency:1,_)-24                              287kB ± 0%     295kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:1000,_sorted:false,_concurrency:2,_)-24                              575kB ± 0%     582kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:1000,_sorted:false,_concurrency:4,_)-24                              533kB ± 0%     557kB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:100000,_sorted:false,_concurrency:1,_)-24                              32.7MB ± 0%    32.7MB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:100000,_sorted:false,_concurrency:2,_)-24                              51.5MB ± 0%    51.5MB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1,_rows:100000,_sorted:false,_concurrency:4,_)-24                              51.5MB ± 0%    51.5MB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:100000,_sorted:false,_concurrency:1,_)-24                             31.7MB ± 0%    30.6MB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:100000,_sorted:false,_concurrency:2,_)-24                             50.5MB ± 0%    49.4MB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:10,_rows:100000,_sorted:false,_concurrency:4,_)-24                             50.5MB ± 0%    49.7MB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24                           27.6MB ± 0%    26.5MB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:2,_)-24                           46.4MB ± 0%    45.3MB ± 0%   ~     (p=1.000 n=1+1)
WindowRows/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24                           46.4MB ± 0%    45.3MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctions/(func:lag,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24                             27.6MB ± 0%    26.5MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctions/(func:lag,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24                             46.5MB ± 0%    45.4MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24             30.7MB ± 0%    28.1MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:2,_)-24             49.6MB ± 0%    47.0MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:3,_)-24             49.5MB ± 0%    46.9MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:4,_)-24             49.7MB ± 0%    46.9MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:5,_)-24             49.7MB ± 0%    47.0MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:1,_ndv:1000,_rows:100000,_sorted:false,_concurrency:6,_)-24             49.7MB ± 0%    46.9MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:1,_)-24             57.8MB ± 0%    48.7MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:2,_)-24             76.7MB ± 0%    67.5MB ± 0%   ~     (p=1.000 n=1+1)
WindowFunctionsWithFrame/(func:row_number,_aggColType:double,_numFunc:5,_ndv:1000,_rows:100000,_sorted:false,_concurrency:3,_)-24             76.6MB ± 0%    67.6MB ± 0%   ~     (p=1.000 n=1+1)
…